add Datetime + Summary models

+++ SERVER/ITS.Monopattino.Server/ITS.Monopattino.Server.Models/Models/BatteryInfo.cs
@@ -13,7 +13,7 @@ namespace ITS.Monopattino.Server.Models.Models
 
         public BatteryInfo(DetectionInfo detection)
         {
-            BatteryLvl = detection.BatteryLvl;
+            BatteryLvl = (int)detection.BatteryLvl;
         }
     }
 }
